+++ automated_tests/features/steps/pktvisor.py
@@ -1,15 +1,8 @@
-import threading
-
 from utils import random_string, threading_wait_until
 import docker
 from behave import step
-import requests
-import yaml
-from yaml.loader import SafeLoader
-import random
-from policies import *
 from test_config import TestConfig
-from utils import check_port_is_available
+from utils import check_port_is_available, make_get_request
 import threading
 from hamcrest import *
 
@@ -29,68 +22,14 @@ def run_pktvisor(context, status_port, role):
     event.wait(1)
 
 
-
 @step("that a pktvisor instance is running on port {status_port} with {role} permission")
 def pkt_running(context, status_port, role):
     run_pktvisor(context, status_port, role)
 
 
-@step("{amount_of_policies} policies {status_condition} be running")
-def amount_of_policies_per_status(context, amount_of_policies, status_condition):
-    assert_that(status_condition, any_of(equal_to("must"), equal_to("must not")),
-                'Unexpect condition for policy status')
-    status_condition_dict = {"must": True, "must not": False}
-    all_policies = make_get_request('policies', context.pkt_port).json()
-    amount_of_policies_per_status = list()
-    for key, value in all_policies.items():
-        if value['input'][list(value['input'].keys())[0]]['input']['running'] is status_condition_dict[
-            status_condition]:
-            amount_of_policies_per_status.append(key)
-    assert_that(len(amount_of_policies_per_status), equal_to(int(amount_of_policies)),
-                f"Unexpect amount of policies that {status_condition} be running")
-
-
-@step("create a new policy with {handler} handler(s)")
-def create_new_policy(context, handler, **kwargs):
-    policy_yaml = policies.generate_policy(handler, random_string(10))
-    policy_yaml_parsed = yaml.load(policy_yaml, Loader=SafeLoader)
-    yaml_policy_data = yaml.dump(policy_yaml_parsed)
-    if "pkt_port" and "expected_status_code" in kwargs:
-        context.response = assert_policy_creation(yaml_policy_data, kwargs["pkt_port"], kwargs["expected_status_code"])
-    elif "pkt_port" in kwargs:
-        context.response = assert_policy_creation(yaml_policy_data, kwargs["pkt_port"])
-    elif "expected_status_code" in kwargs:
-        context.response = assert_policy_creation(yaml_policy_data, context.pkt_port, kwargs["expected_status_code"])
-    else:
-        context.response = assert_policy_creation(yaml_policy_data, context.pkt_port)
-
-
-@step("try to create a new policy with {handler} handler(s)")
-def try_to_create_new_policy(context, handler):
-    create_new_policy(context, handler, pkt_port=context.pkt_port, expected_status_code=404)
-
-
 @step("status code returned on response must be {status_code}")
 def check_response_status_code(context, status_code):
     assert_that(context.response.status_code, equal_to(int(status_code)), "Wrong status code on response")
-
-
-@step("delete {amount_of_policies} policies")
-def remove_policies(context, amount_of_policies):
-    names_of_all_policies = make_get_request('policies', context.pkt_port).json().keys()
-    policies_to_remove = random.sample(names_of_all_policies, int(amount_of_policies))
-    for policy in policies_to_remove:
-        remove_policy(policy, context.pkt_port)
-        response = get_policy(policy, 10853, 404)
-        assert_that(response.json(), has_key('error'), "Unexpected message for non existing policy")
-        assert_that(response.json(), has_value('policy does not exists'), "Unexpected message for non existing policy")
-
-
-@step("try to delete a policy")
-def try_to_delete_policies(context):
-    names_of_all_policies = make_get_request('policies', context.pkt_port).json().keys()
-    sample_policy = random.sample(names_of_all_policies, 1)[0]
-    context.response = remove_policy(sample_policy, context.pkt_port, 404)
 
 
 @step("the pktvisor container status must be {pkt_status}")
@@ -136,39 +75,6 @@ def check_pkt_base_API(context):
         make_get_request(endpoint, context.pkt_port)
 
 
-def assert_policy_creation(yaml_data, pkt_port=10853, expected_status_code=200): #todo arrumar status code
-    """
-
-    :param yaml_data: policy configurations
-    :param pkt_port: port on which pktvisor is running
-    :param expected_status_code: expected status from response
-    :return: response
-    """
-    response = create_policy(yaml_data, pkt_port, expected_status_code)
-    assert_that(response.status_code, equal_to(int(expected_status_code)), f"Post request to create a policy failed "
-                                                                           f"with status {response.status_code}")
-    return response
-
-
-@threading_wait_until
-def make_get_request(end_point, pkt_port=10853, expected_status_code=200, event=None):
-    """
-
-    :param end_point: endpoint to which the request must be sent
-    :param pkt_port: port on which pktvisor is running
-    :param expected_status_code: expected status from response
-    :return: response
-    """
-    pkt_base_api = 'http://localhost:' + str(pkt_port) + '/api/v1/'
-    path = pkt_base_api + end_point
-    response = requests.get(path)
-    if response.status_code == int(expected_status_code):
-        event.set()
-    assert_that(response.status_code, equal_to(int(expected_status_code)),
-                f"Get request to endpoint {path} failed with status {response.status_code}")
-    return response
-
-
 def run_pktvisor_container(container_image, port=10853, role="user", container_name=PKTVISOR_CONTAINER_NAME):
     """
     Run a pktvisor container
@@ -194,34 +100,6 @@ def run_pktvisor_container(container_image, port=10853, role="user", container_n
     return container.id
 
 
-def get_policy(policy_name, pkt_port=10853, expected_status_code=200):
-    """
-
-    :param (str) policy_name: name of the policy to be fetched
-    :param pkt_port: port on which pktvisor is running
-    :param expected_status_code: expected status from response
-    :return: (dict) referred policy data
-    """
-    endpoint = f"policies/{policy_name}"
-    return make_get_request(endpoint, pkt_port, expected_status_code)
-
-
-def remove_policy(policy_name, pkt_port=10853, expected_status_code=200): #todo arrumar status code
-
-    """
-    :param (str) policy_name: name of the policy to be fetched
-    :param pkt_port: port on which pktvisor is . Default: 10853
-    :param expected_status_code: expected status from response. Default: 204
-    :return: response
-    """
-    pkt_base_api = 'http://localhost:' + str(pkt_port) + '/api/v1/policies/'
-    path = pkt_base_api + policy_name
-    response = requests.delete(path)
-    assert_that(response.status_code, equal_to(int(expected_status_code)),
-                f"Delete request of policy {policy_name} failed with status {response.status_code}")
-    return response
-
-
 @threading_wait_until
 def check_amount_of_pkt_with_status(amount_of_pktvisor, pkt_status, event=None):
     docker_client = docker.from_env()
@@ -237,23 +115,3 @@ def check_amount_of_pkt_with_status(amount_of_pktvisor, pkt_status, event=None):
                 event.set()
                 return containers_with_expected_status
     return containers_with_expected_status
-
-
-@threading_wait_until
-def create_policy(yaml_data, pkt_port=10853, expected_status_code=201, event=None):
-    """
-
-    :param yaml_data: policy configurations
-    :param pkt_port: port on which pktvisor is running
-    :param expected_status_code: expected status from response
-    :param event: threading.event
-    :return: response
-    """
-    event.wait(0.5)
-    pkt_api = f"http://localhost:{pkt_port}/api/v1/policies"
-    headers_request = {'Content-type': 'application/x-yaml'}
-    response = requests.post(pkt_api, data=yaml_data, headers=headers_request)
-    if response.status_code == int(expected_status_code):
-        event.set()
-        return response
-    return response
